Reorganise UI code Fixes #211
